Refactored Weather.java
Moved Weather and Powerups to Interactables, as they are not proper actors.
