+++ core/src/com/team21direction/pirategame/Interactables/Weather.java
@@ -1,10 +1,11 @@
-package com.team21direction.pirategame.actors;
+package com.team21direction.pirategame.Interactables;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.scenes.scene2d.Actor;
+import com.team21direction.pirategame.actors.GameActor;
 import com.team21direction.pirategame.screens.MainScreen;
 
 public class Weather extends GameActor {
